+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/jws/SdJwtSerializationTest.kt
@@ -21,12 +21,30 @@ class SdJwtSerializationTest : FreeSpec({
         val value = Random.nextBytes(16).encodeToString(Base64())
         val item = SelectiveDisclosureItem(salt, name, value)
 
-        val serialized = item.serialize().also { println(it) }
+        val serialized = item.serialize()
 
         serialized shouldContain "["
         serialized shouldContain """"${salt.encodeToString(Base64UrlStrict)}""""
         serialized shouldContain """"$name""""
         serialized shouldContain """"$value""""
+        serialized shouldContain "]"
+
+        val deserialized = SelectiveDisclosureItem.deserialize(serialized).getOrThrow()
+        deserialized shouldBe item
+    }
+
+    "Serialization is correct for ByteArray" {
+        val salt = Random.nextBytes(32)
+        val name = Random.nextBytes(16).encodeToString(Base64())
+        val value = Random.nextBytes(16)
+        val item = SelectiveDisclosureItem(salt, name, value)
+
+        val serialized = item.serialize()
+
+        serialized shouldContain "["
+        serialized shouldContain """"${salt.encodeToString(Base64UrlStrict)}""""
+        serialized shouldContain """"$name""""
+        serialized shouldContain """"${value.encodeToString(Base64UrlStrict)}""""
         serialized shouldContain "]"
 
         val deserialized = SelectiveDisclosureItem.deserialize(serialized).getOrThrow()
@@ -39,7 +57,7 @@ class SdJwtSerializationTest : FreeSpec({
         val value = true
         val item = SelectiveDisclosureItem(salt, name, value)
 
-        val serialized = item.serialize().also { println(it) }
+        val serialized = item.serialize()
 
         serialized shouldContain "["
         serialized shouldContain """$value"""
@@ -56,7 +74,7 @@ class SdJwtSerializationTest : FreeSpec({
         val value = Random.nextLong()
         val item = SelectiveDisclosureItem(salt, name, value)
 
-        val serialized = item.serialize().also { println(it) }
+        val serialized = item.serialize()
 
         serialized shouldContain "["
         serialized shouldContain """$value"""
@@ -73,7 +91,7 @@ class SdJwtSerializationTest : FreeSpec({
         val value = Random.nextUInt()
         val item = SelectiveDisclosureItem(salt, name, value)
 
-        val serialized = item.serialize().also { println(it) }
+        val serialized = item.serialize()
 
         serialized shouldContain "["
         serialized shouldContain """$value"""